Massive refactor with unsub logic and lifecycle logic (use #key and abuse return values for $effects)
