--- src/routes/[user]/components/Archive/ListenToDoc.svelte
@@ -1,25 +1,0 @@
-<script>
-  export let theDoc = null
-  export let docPath = ''
-
-  import { doc, onSnapshot } from 'firebase/firestore'
-  import { db } from '/src/lib/db/init.js'
-
-  let unsubListener = null
-
-  $: if (docPath) {
-    destroyAndCreateListener()
-  }
-
-  function destroyAndCreateListener () {
-    theDoc = null
-    if (unsubListener) unsubListener()
-    unsubListener = onSnapshot(doc(db, docPath), doc => {
-      theDoc = { ...doc.data(), id: doc.id }
-    })
-  }
-</script>
-
-<slot theDoc={theDoc}>
-
-</slot>
